requirements.txt
tweepy~=4.4.0
discord~=1.7.3
python-dotenv~=0.19.2